Don't use shared bind mounts
Currently podman is invoked with options to mount both the workspace
directory and some of the files it contains. For example, if the
workspace is `/root`:

```
-v /root/.netrc:/root/.netrc:ro
-v /root:/root:rw,shared
```

The workspace directory is mounted using the _shared_ option. This
results in leaking mount points at an exponential rate, as descrived in
this _podman_ issue (no closed, as it is the expected behaviour):

  https://github.com/containers/podman/issues/15671

This _shared_ option is not really needed. This patch removes it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>