Support wheels distribution
Seems like the comment about data files not being included in wheels
when using full paths is no longer true.

The check I've done is making sure the completion output still works
when installing the wheel files, which seems to work.

In addition, it produces warnings that the legacy installation method
will get deprecated soon:
```
DEPRECATION: strato-skipper was installed using the legacy 'setup.py
install' method, because a wheel could not be built for it. pip 23.1
will enforce this behaviour change.
```